Split into custom hook
